+++ shims/pom.xml
@@ -32,15 +32,6 @@
     <description>The RAPIDS SQL plugin for Apache Spark Shims</description>
     <version>0.2.0-SNAPSHOT</version>
 
-    <profiles>
-        <profile>
-            <id>databricks</id>
-            <modules>
-                <module>databricks</module>
-            </modules>
-        </profile>
-    </profiles>
-
     <modules>
         <module>spark30</module>
         <module>spark31</module>
