+++ shims/spark30/src/main/scala/com/nvidia/spark/rapids/shims/spark30/GpuLast.scala
@@ -18,7 +18,6 @@ package com.nvidia.spark.rapids.shims.spark30
 
 import org.apache.spark.sql.rapids.GpuLastBase
 import com.nvidia.spark.rapids._
-import com.nvidia.spark.rapids.GpuMetricNames._
 
 import org.apache.spark.sql.catalyst.analysis.TypeCheckResult
 import org.apache.spark.sql.catalyst.analysis.TypeCheckResult.{TypeCheckFailure, TypeCheckSuccess}
