+++ shims/spark31/pom.xml
@@ -27,15 +27,9 @@
     </parent>
     <groupId>com.nvidia</groupId>
     <artifactId>rapids-4-spark-shims-spark31_2.12</artifactId>
-    <name>RAPIDS Accelerator for Apache Spark SQL Plugin</name>
-    <description>The RAPIDS SQL plugin for Apache Spark Shim</description>
+    <name>RAPIDS Accelerator for Apache Spark SQL Plugin Spark 3.1 Shim</name>
+    <description>The RAPIDS SQL plugin for Apache Spark 3.1 Shim</description>
     <version>0.2.0-SNAPSHOT</version>
-
-    <profiles>
-        <profile>
-            <id>databricks</id>
-        </profile>
-    </profiles>
 
     <properties>
         <spark.version>3.1.0-SNAPSHOT</spark.version>
