+++ shims/spark31/src/main/scala/com/nvidia/spark/rapids/shims/spark31/GpuBroadcastNestedLoopJoinExec.scala
@@ -14,37 +14,25 @@
  * limitations under the License.
  */
 
-package com.nvidia.spark.rapids.shims
+package com.nvidia.spark.rapids.shims.spark31
 
 import com.nvidia.spark.rapids._
-import com.nvidia.spark.rapids.GpuMetricNames._
+
+import org.apache.spark.internal.Logging
+import org.apache.spark.sql.catalyst.expressions.Expression
+import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight}
+import org.apache.spark.sql.catalyst.plans.JoinType
+import org.apache.spark.sql.execution.SparkPlan
+import org.apache.spark.sql.execution.joins.BroadcastNestedLoopJoinExec
 import org.apache.spark.sql.rapids.execution._
 
-import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight, BuildSide}
-import org.apache.spark.internal.Logging
-import org.apache.spark.sql.execution.joins.BroadcastNestedLoopJoinExec
-import org.apache.spark.sql.execution.joins.ShuffledHashJoinExec
-import org.apache.spark.TaskContext
-import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.catalyst.InternalRow
-import org.apache.spark.sql.catalyst.expressions.Expression
-import org.apache.spark.sql.catalyst.plans.JoinType
-import org.apache.spark.sql.catalyst.plans.physical.{Distribution, HashClusteredDistribution}
-import org.apache.spark.sql.execution.{BinaryExecNode, SparkPlan}
-import org.apache.spark.sql.execution.joins._
-import org.apache.spark.sql.execution.metric.{SQLMetric, SQLMetrics}
-import org.apache.spark.sql.vectorized.ColumnarBatch
-import org.apache.spark.internal.Logging
-
-
-
-case class GpuBroadcastNestedLoopJoinExec31(
+case class GpuBroadcastNestedLoopJoinExec(
     left: SparkPlan,
     right: SparkPlan,
     join: BroadcastNestedLoopJoinExec,
     joinType: JoinType,
-    condition: Option[Expression]) extends GpuBroadcastNestedLoopJoinExecBase(left, right, join, joinType, condition) with Logging {
-
+    condition: Option[Expression])
+  extends GpuBroadcastNestedLoopJoinExecBase(left, right, join, joinType, condition) with Logging {
 
   def getBuildSide: GpuBuildSide = {
     join.buildSide match {
@@ -55,7 +43,7 @@ case class GpuBroadcastNestedLoopJoinExec31(
   }
 }
 
-object GpuBroadcastNestedLoopJoinExec31 extends Logging {
+object GpuBroadcastNestedLoopJoinExec extends Logging {
 
   def createInstance(
       left: SparkPlan,
@@ -64,7 +52,6 @@ object GpuBroadcastNestedLoopJoinExec31 extends Logging {
       joinType: JoinType,
       condition: Option[Expression]): GpuBroadcastNestedLoopJoinExecBase = {
     
-    GpuBroadcastNestedLoopJoinExec31(left, right, join, joinType, condition)
+    GpuBroadcastNestedLoopJoinExec(left, right, join, joinType, condition)
   }
-
 }
