+++ shims/spark31/src/main/scala/com/nvidia/spark/rapids/shims/spark31/GpuLast.scala
@@ -17,8 +17,6 @@
 package com.nvidia.spark.rapids.shims.spark31
 
 import org.apache.spark.sql.rapids.GpuLastBase
-import com.nvidia.spark.rapids._
-import com.nvidia.spark.rapids.GpuMetricNames._
 
 import org.apache.spark.sql.catalyst.analysis.TypeCheckResult
 import org.apache.spark.sql.catalyst.analysis.TypeCheckResult.{TypeCheckFailure, TypeCheckSuccess}
