+++ shims/spark31/src/main/scala/com/nvidia/spark/rapids/shims/spark31/Spark31Shims.scala
@@ -19,9 +19,9 @@ package com.nvidia.spark.rapids.shims.spark31
 import java.time.ZoneId
 
 import com.nvidia.spark.rapids._
+import com.nvidia.spark.rapids.shims.GpuBroadcastNestedLoopJoinExec31
 import org.apache.spark.sql.rapids.GpuTimeSub
 import org.apache.spark.sql.rapids.shims.spark31._
-
 import org.apache.spark.sql.catalyst.expressions.aggregate.{First, Last}
 import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight, BuildSide}
 import org.apache.spark.internal.Logging
@@ -30,11 +30,23 @@ import org.apache.spark.sql.execution.datasources.HadoopFsRelation
 import org.apache.spark.sql.execution.joins.{BroadcastHashJoinExec, BroadcastNestedLoopJoinExec, SortMergeJoinExec}
 import org.apache.spark.sql.execution.joins.ShuffledHashJoinExec
 import org.apache.spark.sql.catalyst.expressions._
+import org.apache.spark.sql.catalyst.plans.JoinType
+import org.apache.spark.sql.rapids.execution.GpuBroadcastNestedLoopJoinExecBase
 import org.apache.spark.unsafe.types.{CalendarInterval, UTF8String}
 import org.apache.spark.sql.types._
 
 
 class Spark31Shims extends SparkShims with Logging {
+
+  def getGpuBroadcastNestedLoopJoinShims(
+      left: SparkPlan,
+      right: SparkPlan,
+      join: BroadcastNestedLoopJoinExec,
+      joinType: JoinType,
+      condition: Option[Expression]): GpuBroadcastNestedLoopJoinExecBase = {
+    GpuBroadcastNestedLoopJoinExec(left, right, join, joinType, condition)
+  }
+
 
   def isGpuHashJoin(plan: SparkPlan): Boolean = {
     plan match {
