+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/SparkShims.scala
@@ -16,23 +16,17 @@
 
 package com.nvidia.spark.rapids
 
-import org.apache.spark.TaskContext
-import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.catalyst.plans.JoinType
-import org.apache.spark.sql.catalyst.plans.physical.{Distribution, HashClusteredDistribution}
-import org.apache.spark.sql.execution.{BinaryExecNode, SparkPlan}
+import org.apache.spark.sql.execution.SparkPlan
 import org.apache.spark.sql.execution.joins._
-import org.apache.spark.sql.execution.metric.{SQLMetric, SQLMetrics}
-import org.apache.spark.sql.vectorized.ColumnarBatch
+import org.apache.spark.sql.rapids.execution.GpuBroadcastNestedLoopJoinExecBase
 
 sealed abstract class GpuBuildSide
 
 case object GpuBuildRight extends GpuBuildSide
 
 case object GpuBuildLeft extends GpuBuildSide
-
 
 trait SparkShims {
 
@@ -42,7 +36,12 @@ trait SparkShims {
   def getBuildSide(join: BroadcastHashJoinExec): GpuBuildSide
   def getExprs: Seq[ExprRule[_ <: Expression]]
   def getExecs: Seq[ExecRule[_ <: SparkPlan]]
-
+  def getGpuBroadcastNestedLoopJoinShims(
+    left: SparkPlan,
+    right: SparkPlan,
+    join: BroadcastNestedLoopJoinExec,
+    joinType: JoinType,
+    condition: Option[Expression]): GpuBroadcastNestedLoopJoinExecBase
 }
